abstract partition clustering class creation
